pyproject.toml
[project]
name = "sinengus"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "aiogram (>=3.19.0,<4.0.0)",
    "transformers (>=4.49.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
